backend/requirements.txt
langchain>=0.1.14
groq>=0.4.0
streamlit>=1.33.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
chromadb>=0.4.22
sentence-transformers>=2.6.1
python-dotenv>=1.0.1
langchain_community
PyPDF2>=3.0.1
fpdf>=1.7.2
requests>=2.31.0
gtts>=2.5.1
streamlit-webrtc>=0.45.0
av>=10.0.0
SpeechRecognition>=3.10.0
numpy>=1.24.0
